More refactoring of Tree
Visitor: new class, split off from Tree;
Tree: implement map method.
